pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "inclusive-job-ad-analyser"
version = "1.0.0"
description = "NLP-powered tool for detecting biased language in job descriptions"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["nlp", "bias-detection", "inclusive-hiring", "job-ads", "ethics"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Human Resources",
    "Topic :: Text Processing :: Linguistic",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "pandas>=2.0.0,<3.0.0",
    "PyYAML>=6.0,<7.0",
    "colorama>=0.4.6,<1.0.0",
    "spacy>=3.7.0,<4.0.0",
]

[project.optional-dependencies]
web = ["gradio>=4.0.0,<5.0.0"]
dev = [
    "pytest>=7.4.0,<8.0.0",
    "pytest-cov>=4.1.0,<5.0.0",
    "black>=23.0.0,<24.0.0",
    "flake8>=6.0.0,<7.0.0",
    "mypy>=1.5.0,<2.0.0",
]

[project.scripts]
job-ad-analyser = "inclusive_job_ad_analyser.cli:main"

[project.urls]
Homepage = "https://github.com/okuye/inclusive-job-ad-analyser"
Repository = "https://github.com/okuye/inclusive-job-ad-analyser"
Issues = "https://github.com/okuye/inclusive-job-ad-analyser/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=src --cov-report=html --cov-report=term-missing"

[tool.black]
line-length = 88
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
